feat: only user can update and delete event

+++ src/events/events.controller.ts
@@ -2,9 +2,9 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
-  // Logger,
   NotFoundException,
   Param,
   ParseIntPipe,
@@ -15,11 +15,7 @@ import {
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { MoreThan, Repository } from 'typeorm';
-import { Attendee } from './attendee.entity';
 import { CreateEventDto } from './inputs/create-event.dto';
-import { Event } from './event.entity';
 import { EventsService } from './events.service';
 import { UpdateEventDto } from './inputs/update-event.dto';
 import { ListEvents } from './inputs/list.events';
@@ -31,13 +27,7 @@ import { AuthGuardJwt } from '../auth/auth-guard.jwt';
 export class EventsController {
   // readonly #logger = new Logger(EventsController.name);
 
-  constructor(
-    @InjectRepository(Event)
-    private readonly eventRepository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeeRepository: Repository<Attendee>,
-    private readonly eventsService: EventsService,
-  ) {}
+  constructor(private readonly eventsService: EventsService) {}
 
   @Get()
   @UsePipes(new ValidationPipe({ transform: true }))
@@ -52,38 +42,38 @@ export class EventsController {
     );
   }
 
-  @Get('/practice')
-  async Practice() {
-    return await this.eventRepository.find({ where: { id: MoreThan(2) } });
-  }
+  // @Get('/practice')
+  // async Practice() {
+  //   return await this.eventRepository.find({ where: { id: MoreThan(2) } });
+  // }
 
-  @Get('/practice2')
-  async Practice2() {
-    // const event = await this.eventRepository.findOne({
-    //   where: { id: 1 },
-    //   relations: { attendees: true },
-    // });
-    // if (!event) {
-    //   throw new NotFoundException();
-    // }
-    // const attendee = new Attendee();
-    // attendee.name = 'Jerry';
-    // attendee.event = event;
-    // await this.attendeeRepository.save({ ...attendee });
-
-    // With Cascade - set cascade to true in oneToMany relation to event entity
-    // const attendee = new Attendee();
-    // attendee.name = 'Jerry with cascade';
-    // event.attendees.push(attendee);
-    // await this.eventRepository.save({ ...event });
-
-    return await this.eventRepository
-      .createQueryBuilder('e')
-      .select(['e.id', 'e.name'])
-      .orderBy('e.id', 'ASC')
-      .take(3)
-      .getMany();
-  }
+  // @Get('/practice2')
+  // async Practice2() {
+  //   // const event = await this.eventRepository.findOne({
+  //   //   where: { id: 1 },
+  //   //   relations: { attendees: true },
+  //   // });
+  //   // if (!event) {
+  //   //   throw new NotFoundException();
+  //   // }
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'Jerry';
+  //   // attendee.event = event;
+  //   // await this.attendeeRepository.save({ ...attendee });
+  //
+  //   // With Cascade - set cascade to true in oneToMany relation to event entity
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'Jerry with cascade';
+  //   // event.attendees.push(attendee);
+  //   // await this.eventRepository.save({ ...event });
+  //
+  //   return await this.eventRepository
+  //     .createQueryBuilder('e')
+  //     .select(['e.id', 'e.name'])
+  //     .orderBy('e.id', 'ASC')
+  //     .take(3)
+  //     .getMany();
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -105,22 +95,34 @@ export class EventsController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.eventRepository.findOneBy({ id });
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventsService.getEvent(id);
     if (!event) {
       throw new NotFoundException();
     }
-    return await this.eventRepository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+    if (event.organizerId != user.id) {
+      throw new ForbiddenException(null, 'You are not allowed to change this');
+    }
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id);
-    if (result?.affected !== 1) throw new NotFoundException();
+  async remove(@Param('id') id, @CurrentUser() user: User) {
+    const event = await this.eventsService.getEvent(id);
+
+    if (!event) {
+      throw new NotFoundException();
+    }
+    if (event.organizerId != user.id) {
+      throw new ForbiddenException(null, 'You are not allowed to remove this');
+    }
+    await this.eventsService.deleteEvent(id);
   }
 }
